+++ assets/i18n/en.json
@@ -21,6 +21,7 @@
     "ENTER_MESSAGE_HERE": "Enter Message Here",
     "CREATE": "Create",
     "SEARCH": "Search",
+    "SEARCH_RECORDS": "Search records",
     "STATUS": "Status"
   },
   "NEW_INTERVENTION": {
