+++ assets/i18n/ro.json
@@ -21,6 +21,7 @@
     "ENTER_MESSAGE_HERE": "Introduceti Mesajul Aici",
     "CREATE": "Creaza",
     "SEARCH": "Cauta",
+    "SEARCH_RECORDS": "Cauta inregistrari",
     "STATUS": "Status"
   },
   "NEW_INTERVENTION": {
