+++ package.json
@@ -70,8 +70,8 @@
     "@unicef-polymer/etools-dexie-caching": "^1.0.2",
     "@unicef-polymer/etools-dialog": "6.0.0",
     "@unicef-polymer/etools-dropdown": "^3.7.9",
+    "@unicef-polymer/etools-modules-common": "1.1.0-rc.1",
     "@unicef-polymer/etools-loading": "5.0.0",
-    "@unicef-polymer/etools-modules-common": "1.0.8",
     "@unicef-polymer/etools-profile-dropdown": "7.0.0",
     "@unicef-polymer/etools-info-tooltip": "4.1.2",
     "@unicef-polymer/etools-table": "^1.3.0",
