+++ src_ts/app-shell.ts
@@ -95,6 +95,8 @@ import {logError} from '@unicef-polymer/etools-behaviors/etools-logging';
 import get from 'lodash-es/get';
 import {EtoolsRouter} from './components/utils/routes.js';
 import {registerTranslateConfig, use} from 'lit-translate';
+import {getRedirectToListPath} from './components/utils/subpage-redirect';
+import debounce from 'lodash-es/debounce';
 declare const dayjs: any;
 declare const dayjs_plugin_utc: any;
 declare const dayjs_plugin_isSameOrBefore: any;
@@ -351,6 +353,7 @@ class AppShell extends connect(store)(
   }
 
   public connectedCallback() {
+    this.updateReduxRouteDetails = debounce(this.updateReduxRouteDetails.bind(this), 20);
     super.connectedCallback();
 
     this.checkAppVersion();
@@ -378,6 +381,11 @@ class AppShell extends connect(store)(
 
   updateReduxRouteDetails(appLocRoute: any) {
     const routeDetails = EtoolsRouter.getRouteDetails(appLocRoute);
+    // If the url is not complete(ex /pmp/interventions), redirect to /pmp/interventions/list
+    const redirectTo = getRedirectToListPath(appLocRoute.path);
+    if (redirectTo) {
+      EtoolsRouter.replaceAppLocation(redirectTo);
+    }
     if (!isJsonStrMatch(routeDetails, get(store.getState(), 'app.routeDetails'))) {
       store.dispatch(updateStoreRouteDetails(routeDetails));
     }
