+++ src_ts/components/pages/partners/pages/financial-assurance/components/hact-edit-dialog.ts
@@ -3,17 +3,18 @@ import {fireEvent} from '../../../../../utils/fire-custom-event';
 import EtoolsDialog from '@unicef-polymer/etools-dialog/etools-dialog';
 import '@polymer/paper-input/paper-input';
 import '@unicef-polymer/etools-dropdown/etools-dropdown-multi';
-import EndpointsMixin from '../../../../../endpoints/endpoints-mixin-lit';
 import {gridLayoutStylesLit} from '@unicef-polymer/etools-modules-common/dist/styles/grid-layout-styles-lit';
 import clone from 'lodash-es/clone';
 import {sendRequest} from '@unicef-polymer/etools-ajax/etools-ajax-request';
 import {parseRequestErrorsAndShowAsToastMsgs} from '@unicef-polymer/etools-ajax/ajax-error-parser';
 import {GenericObject} from '@unicef-polymer/etools-types';
-import CommonMixin from '../../../../../common/mixins/common-mixin-lit';
+import CommonMixinLit from '../../../../../common/mixins/common-mixin-lit';
 import {translate} from 'lit-translate';
+import EndpointsLitMixin from '@unicef-polymer/etools-modules-common/dist/mixins/endpoints-mixin-lit';
+import pmpEdpoints from '../../../../../endpoints/endpoints';
 
 @customElement('hact-edit-dialog')
-export class HactEditDialog extends CommonMixin(EndpointsMixin(LitElement)) {
+export class HactEditDialog extends CommonMixinLit(EndpointsLitMixin(LitElement)) {
   static get styles() {
     return [gridLayoutStylesLit];
   }
@@ -308,7 +309,7 @@ export class HactEditDialog extends CommonMixin(EndpointsMixin(LitElement)) {
   savePartnerDetails(body: any, partnerId: string) {
     const params = {
       method: 'PATCH',
-      endpoint: this.getEndpoint('partnerDetails', {id: partnerId}),
+      endpoint: this.getEndpoint(pmpEdpoints, 'partnerDetails', {id: partnerId}),
       body
     };
     sendRequest(params)
